% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcdfrankenstein.R
\name{frankenstein}
\alias{frankenstein}
\title{Tool to help plot people}
\usage{
frankenstein(graph, xrange, yrange, ...)
}
\arguments{
\item{graph}{This is graph that will be use to as the back 
drop for the xkcdman}

\item{xrange}{xkcdman needs a xrange plus helpful for the tool it self}

\item{yrange}{xkcdman needs a yrange}

\item{...}{Other arguments}
}
\value{
Launching shiny app
}
\description{
Let's face it, plotting the xkcdman is hard, so I created 
a shiny app called Frankenstein to help with this process.
}
\examples{
\dontrun{
xrange <- range(mtcars$mpg)
yrange <- range(mtcars$wt)
p <- ggplot() +
     geom_point(aes(mpg, wt), data=mtcars) +
     xkcdaxis(xrange, yrange)
frankenstein(p, xrange, yrange) # launches app
}
}
