% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcd.R
\name{init}
\alias{init}
\alias{x_chart}
\title{Initialize}
\usage{
x_chart(data, title = NULL, ..., width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{data frame}

\item{title}{chart title}

\item{...}{other configuration options, see \code{\link{x_options}}}
}
\description{
for specific chart types, either one of \code{x_line}, \code{x_point}, \code{x_bar}, \code{x_pie} or \code{x_radar} needs to be appended after \code{x_chart}
}
\details{
Initialize a xkcd widget
}
