% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcd.R
\name{plot.xkcd}
\alias{plot.xkcd}
\title{Visualize xkcd comics}
\usage{
\method{plot}{xkcd}(x, ...)
}
\arguments{
\item{x}{an \code{\link{xkcd}} object}

\item{...}{currently ignored}
}
\description{
Given an \code{\link{xkcd}} object, this \code{\link[base:plot]{base::plot}} method retrieves the image file associated with
this comic from the xkcd website and displays it in the currently active graphics device.
}
\examples{
\dontrun{
first_comic <- xkcd(1)
plot(first_comic)
}
}
