% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{x_bar}
\alias{x_bar}
\title{Bar, pie and donut charts}
\usage{
x_bar(w, x, y, fill = NULL)
}
\arguments{
\item{w}{an xkcd widget object}

\item{x}{x variable}

\item{y}{y variable for bar height or arc length in pie chart and donut chart}

\item{fill}{optional group variable mapped to bar fill}
}
\description{
Bar, pie and donut charts
}
