% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{x_line}
\alias{x_line}
\title{Line Graph}
\usage{
x_line(w, x = "x", y = "y", color = NULL)
}
\arguments{
\item{w}{an xkcd widget object}

\item{x}{x variable}

\item{y}{y variable}

\item{color}{optional group variable mapped to colors}
}
\description{
Line Graph
}
\examples{
subset(airquality, Month == 5) |>
 x_chart() |>
 x_line(x = "Day", y = "Temp")
}
