% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{x_point}
\alias{x_point}
\title{Scatterplot}
\usage{
x_point(w, x, y, fill = NULL, size = 1, line = FALSE)
}
\arguments{
\item{w}{an xkcd widget object}

\item{x}{x variable}

\item{y}{y variable}

\item{fill}{optional group variable mapped to dot fill}

\item{size}{dot size}

\item{line}{if dots should be connected by line}
}
\description{
Scatterplot
}
