% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar.R
\name{x_radar}
\alias{x_radar}
\title{Radar chart}
\usage{
x_radar(w, x, y, color = NULL, label = TRUE, size = 1, ticks = 3)
}
\arguments{
\item{w}{an xkcd widget object}

\item{x}{x variable (dimension)}

\item{y}{y variable (spoke length)}

\item{color}{optional group variable mapped to colors}

\item{label}{display labels for each dimension}

\item{size}{dot size}

\item{ticks}{number of ticks on the main line}
}
\description{
Radar chart
}
