% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{xkcd-options}
\alias{xkcd-options}
\alias{x_options}
\alias{x_label}
\alias{x_title}
\alias{x_legend}
\title{Common configuration options}
\usage{
x_options(
  w,
  title = NULL,
  xlabel = NULL,
  ylabel = NULL,
  palette = NULL,
  font_family = "xkcd",
  stroke = "black",
  background = "white",
  legend = true,
  legend_position = 1,
  ...
)

x_label(w, title = NULL, xlabel = NULL, ylabel = NULL)

x_title(w, title = NULL)

x_legend(w, legend = TRUE, legend_position = 1)
}
\arguments{
\item{w:}{an xkcd widget object}

\item{title:}{title}

\item{xlabel:}{title for x axis}

\item{ylabel:}{title for y axis}

\item{palette:}{color palette for different groups}

\item{font_family:}{font family}

\item{stroke:}{color for axis, title and legend box}

\item{background:}{background color}

\item{legend:}{if legend should be shown}

\item{legend_position:}{a integer controlling legend position, 1 (up left), 2 (up right), 3 (down left), 4 (down right)}
}
\description{
various functions for controlling guide elements such as title, axis label, legend, and fonts
}
