\name{xkcd-package}
\alias{xkcd-package}
\alias{xkcd}
\docType{package}
\title{
\packageTitle{xkcd}
}
\description{
\packageDescription{xkcd}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{xkcd}
\packageIndices{xkcd}
}
\author{
\packageAuthor{xkcd}

Maintainer: \packageMaintainer{xkcd}
}
\keyword{ package }
\examples{
\dontrun{vignette("xkcd-intro")}
}
