% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcdline.R
\name{xkcdline}
\alias{xkcdline}
\title{Draw lines or circles}
\usage{
xkcdline(mapping, data, typexkcdline = "segment", mask = TRUE, ...)
}
\arguments{
\item{mapping}{Aesthetic mapping}

\item{data}{Dataset}

\item{typexkcdline}{"segment" or "circunference"}

\item{mask}{Logical}

\item{...}{Additional arguments}
}
\description{
Deprecated in favour of [geom_xkcdpath()], which integrates with ggplot2's
aesthetic machinery instead of rebuilding mappings by hand. Migrate segments to
`geom_xkcdpath(aes(x = x, y = y, xend = xend, yend = yend), ...)` and circles to
the `diameter` aesthetic.
}
\seealso{
[geom_xkcdpath()]
}
\keyword{internal}
