% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_xkcd.R
\name{theme_xkcd}
\alias{theme_xkcd}
\title{Creates an XKCD theme}
\usage{
theme_xkcd()
}
\value{
A \code{\link[ggplot2]{theme}} object.
}
\description{
This function creates an XKCD theme, applying the 'xkcd' font if available.
}
\note{
The "xkcd" font must be installed and registered with \code{extrafont} for the
full effect. See the vignette \code{vignette("xkcd-intro")} for installation instructions.
}
\examples{
\dontrun{
# Assuming 'xkcd' font is installed and registered:
p <- ggplot(mtcars, aes(mpg, wt)) +
     geom_point() +
     theme_xkcd()
p
}
}
