% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{x_bar}
\alias{x_bar}
\alias{x_pie}
\title{Bar, pie/doughnut charts}
\usage{
x_bar(w, x, y, fill = NULL)

x_pie(w, x, y, radius = 0.5)
}
\arguments{
\item{w}{an xkcd widget object}

\item{x}{x variable}

\item{y}{y variable for bar height or arc length in pie chart}

\item{fill}{optional group variable mapped to bar fill}

\item{radius}{specify empty pie chart radius, default to 0.5 For pie chart set radius = 0}
}
\description{
Bar, pie/doughnut charts
}
