% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{x_line}
\alias{x_line}
\title{Line Graph}
\usage{
x_line(w, x, y, color = NULL)
}
\arguments{
\item{x:}{x variable}

\item{y:}{y variable}

\item{color:}{optional group variable mapped to colors}
}
\description{
Line Graph
}
