% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xkcd-package.R
\docType{package}
\name{xkcd-package}
\alias{xkcd}
\alias{xkcd-package}
\title{Drawing XKCD graphics}
\description{
XKCD is a popular stick figure web comic  with themes in  mathematics, science, language, and romance created by Randall Munroe. This package gives a satisfactory answer to the question How can we make xkcd style graphs in R?. It provides a set of functions for plotting data in an XKCD style using ggplot2.
}
\details{
\tabular{ll}{
Package: \tab xkcd\cr
Type: \tab Package\cr
Version: \tab 0.0.4\cr
Date: \tab 2014-12-08\cr
License: \tab GPL 3.0\cr }
}
\examples{
\dontrun{vignette("xkcd-intro")}
}
\author{
Emilio Torres-Manzanera

Maintainer: Emilio Torres-Manzanera <torres@uniovi.es>
}
\keyword{package}

