% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcd.R
\name{xkcd}
\alias{xkcd}
\title{xkcd}
\usage{
xkcd(number)
}
\arguments{
\item{number}{A scalar numeric vector identifying  the xkcd comic}
}
\value{
a list of class 'xkcd' with the following fields
\itemize{
\item month
\item num
\item link
\item safe_title
\item transcript
\item alt
\item img
\item title
\item day
}

Note that many of these field may be an empty string
}
\description{
Retrieve metadata about and xkcd comic
given an \code{\link{xkcd}} comic number this function retrieves and JSON object
describing that comic from the xkcd website
}
