% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{xkcd}
\alias{xkcd}
\title{xkcd: R interface to chart.xkcd}
\description{
The \code{xkcd} package creates xkcd-styled chart widgets, including line charts, scatterplots,
bar plots, pie chart and radar charts.
}
\seealso{
official documentation for chart.xkcd library: \url{https://timqian.com/chart.xkcd/}
}
