% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcd.R
\name{xkcd}
\alias{xkcd}
\title{use the xkcd API}
\usage{
xkcd(number)
}
\description{
retrieves info about the xkcd comic strips, given a comic number, calls the xkcd JSON API, returns metadata about the comic in the format of a list
}
