% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcd.R
\name{xkcd}
\alias{xkcd}
\title{Retrieve metadata about an xkcd comic}
\usage{
xkcd(number)
}
\arguments{
\item{number}{A scalar numeric vector identifying an xkcd comic number. The numeric value is treated as an integer, and any fractional component is always rounded down.}
}
\value{
description A list with the following fields
\itemize{
\item month
\item num
\item link
\item year
\item news
\item safe_title
\item transcript
\item alt
\item img
\item title
\item day
}

Note that many of these fields may be an empty string
}
\description{
Given an xkcd comic number, this function retrieves a json object describing that comic from the
xkcd website
}
