% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcd.R
\name{xkcd}
\alias{xkcd}
\title{Retrieve metadata about an xkcd comic}
\usage{
xkcd(number)
}
\description{
Given an xkcd comic number, this function retrieves a JSON object describing that comic from the official xkcd API.
}
