% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcd.R
\name{xkcd}
\alias{xkcd}
\title{Use the xkcd API}
\usage{
xkcd(number)
}
\arguments{
\item{number}{A scalar numeric vector corresponding to an xkcd comic number}
}
\value{
A list of 11 elements
\itemize{
\item \code{month}: A scalar character vector
\item \code{num}: A scalar numeric vector
\item \code{link}: A scalar character vector, may be empty
\item \code{year}: A scalar character vector
\item \code{news}: A scalar character vector, maybe empty
\item \code{safe_title}: A scalar character vector
\item \code{transcript}: A scalar character vector
\item \code{alt}: A scalar character vector
\item \code{img}: A scalar character vector
\item \code{title}: A scalar character vector
\item \code{day}: A scalar character vector
}
}
\description{
Given a comic number, the \code{xkcd()} function calls the xkcd JSON API and returns metadata about the comic in the form of a list object.
}
\examples{
dino_time <- xkcd(15)
str(dino_time)
}
