% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcd.R
\name{xkcd}
\alias{xkcd}
\title{retrieve metadata about and xkcd comic}
\usage{
xkcd(number)
}
\description{
Given an xkcd comic number, this function retrieves a JSON objevt describing that comic from the official xkdc API.
}
