% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcdline.R
\name{xkcdline}
\alias{xkcdline}
\title{Draw lines or circles in an XKCD style (Deprecated)}
\usage{
xkcdline(mapping = NULL, data = NULL, typexkcdline = "segment", ...)
}
\arguments{
\item{mapping}{Aesthetic mapping generated by \code{\link[ggplot2]{aes}}.}

\item{data}{Dataset used in this layer.}

\item{typexkcdline}{DEPRECATED. The new Geom detects type from aesthetics.}

\item{...}{Other arguments passed to the Geom.}
}
\description{
This function is deprecated. Use \code{\link{geom_xkcdpath}} instead.
}
