% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcdline.R
\name{xkcdline}
\alias{xkcdline}
\title{Draw lines or circles}
\usage{
xkcdline(mapping, data, typexkcdline = "segment", mask = TRUE, ...)
}
\arguments{
\item{mapping}{Aesthetic mapping}

\item{data}{Dataset}

\item{typexkcdline}{"segment" or "circunference"}

\item{mask}{Logical}

\item{...}{Additional arguments}
}
\description{
Draw lines or circles
}
