% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcdman.R
\name{xkcdman}
\alias{xkcdman}
\title{Draw a stick figure}
\usage{
xkcdman(mapping, data, ...)
}
\arguments{
\item{mapping}{Mapping between variables and aesthetics generated by \code{\link[ggplot2]{aes}}.}

\item{data}{Dataset used in this layer.}

\item{...}{Other arguments passed to geom_xkcdpath.}
}
\value{
A list of layers forming the stick figure.
}
\description{
This function draws an XKCD style stick figure.

The following aesthetics are required:
\itemize{
  \item x: x position of the center of the head.
  \item y: y position of the center of the head.
  \item scale: scale of the man (size in Y axis units).
  \item ratioxy: Ratio x to y of the graph (Use \code{diff(xrange) / diff(yrange)}).
  \item angleofspine: angle between the spine and a horizontal line.
  \item anglerighthumerus, anglelefthumerus, etc.: angles for limbs.
  \item angleofneck: angle for the neck segment.
}
}
\examples{
\dontrun{
man <- data.frame(x=1,y=1,scale=1,ratioxy=1,
                  angleofspine= -90,
                  anglerighthumerus=-120,anglelefthumerus=-60,
                  anglerightradius=-120,angleleftradius=-60,
                  anglerightleg=-140,angleleftleg=-140,
                  angleofneck=-90)
p <- ggplot() + xkcdman(aes(x=x,y=y,scale=scale,ratioxy=ratioxy,
                            angleofspine=angleofspine,
                            anglerighthumerus=anglerighthumerus,
                            anglelefthumerus=anglelefthumerus,
                            anglerightradius=anglerightradius,
                            angleleftradius=angleleftradius,
                            anglerightleg=anglerightleg,
                            angleleftleg=angleleftleg,
                            angleofneck=angleofneck),
                       data=man) + xkcdaxis(c(0,2),c(0,2))
p
}
}
\seealso{
\code{\link[ggplot2]{aes}}, \code{\link{geom_xkcdpath}}
}
