% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcdman.R
\name{xkcdman}
\alias{xkcdman}
\title{Draw a stick figure}
\usage{
xkcdman(mapping, data, ...)
}
\arguments{
\item{mapping}{Aesthetic mapping}

\item{data}{Dataset}

\item{...}{Optional arguments}
}
\description{
Draw a stick figure
}
