% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcdrect.R
\name{xkcdrect}
\alias{xkcdrect}
\title{Draw fuzzy rectangles}
\usage{
xkcdrect(
  mapping,
  data,
  ...,
  fillcolour = "grey90",
  bordercolour = "black",
  borderlinewidth = 0.5,
  borderxjitteramount = 0.005,
  borderyjitteramount = 0.005
)
}
\arguments{
\item{mapping}{Mapping between variables and aesthetics generated by \code{\link[ggplot2]{aes}}. See Details.}

\item{data}{Dataset used in this layer.}

\item{...}{Other arguments passed to geom_xkcdpath.}

\item{fillcolour}{Colour for the fill of the rectangle.}

\item{bordercolour}{Colour for the border of the rectangle.}

\item{borderxjitteramount}{Jitter on the X-axis for vertical borders.}

\item{borderyjitteramount}{Jitter on the Y-axis for horizontal borders.}

\item{bordersize}{Size for the border of the rectangle (passed to size aesthetic).}
}
\value{
A list of layers for the rectangle.
}
\description{
This function draws fuzzy rectangles by plotting four separate, jittered line segments.

It plots rectangles. The following aesthetics are required:
\itemize{
  \item xmin
  \item ymin
  \item xmax
  \item ymax
}
Additionally, you can use the aesthetics of \code{\link[ggplot2]{geom_path}} and \code{\link[ggplot2]{geom_rect}}, as well as parameters for \code{\link{geom_xkcdpath}}.
}
\examples{
\dontrun{
volunteers <- data.frame(year=c(2007:2011),
                        number=c(56470, 56998,59686, 61783, 64251))
p <- ggplot() + xkcdrect(aes(xmin = year,
                             xmax= year +0.3,
                             ymin= number - 500,
                             ymax= number + 500),
                             data=volunteers, fillcolour = "pink")
p
}
}
\seealso{
\code{\link[ggplot2]{aes}},  \code{\link[ggplot2]{geom_path}}
}
